Eliminate the need to route Zootopia through autodj by switching the console feeds
Prior to this change Zootopia was enabled by enabling it to route through autodj
.
Now Zootopia is enabled by calling an ZootipiaOn.applescript which calls a
shell script that then flips the Zootopia feed relay in the JSM console.
Conversely, the new AutodjOn.applecript enables the autodj feed at the console.